refact: abstracted broadcasting of records
