+++ src/DashJS/SDK/Platform/methods/contracts/broadcast.ts
@@ -1,6 +1,3 @@
-import {Platform} from "../../Platform";
-import StateTransitionBuilder, {StateTransitionBuilderTypes} from "../../../StateTransitionBuilder/StateTransitionBuilder";
-
 /**
  * Broadcast contract onto the platform
  *
@@ -8,14 +5,5 @@ import StateTransitionBuilder, {StateTransitionBuilderTypes} from "../../../Stat
  * @param contract - contract
  * @param identity - identity
  */
-export async function broadcast(this: Platform, contract: any, identity: any): Promise<any> {
-    const {account, client, dpp} = this;
-
-    const builder = new StateTransitionBuilder({type: StateTransitionBuilderTypes.CONTRACT, dpp, client});
-    builder.addRecord(contract);
-    // @ts-ignore
-    const identityPrivateKey = account.getIdentityHDKey(0, 'user').privateKey;
-    await builder.register(identity, identityPrivateKey);
-}
-
+import broadcast from "../broadcastRecords";
 export default broadcast;
